backend/requirements.txt
fastapi>=0.135.1
uvicorn>=0.41.0
gunicorn>=22.0.0
treepoem>=3.28.0
sqlalchemy[asyncio]>=2.0.48
asyncpg>=0.30.0
alembic>=1.13.0
pandas>=3.0.1
python-multipart>=0.0.22
openpyxl>=3.1.5
pillow>=10.0.0
python-dotenv>=1.0.0
# Auth
fastapi-users[sqlalchemy]>=13.0.0
PyJWT>=2.8.0
bcrypt>=4.0.0
# Cache / rate limiting
redis>=5.0.0
# Email
resend>=2.0.0
